requirements.txt
torch>=1.12
numpy>=1.23
matplotlib>=3.5.3
opencv-python>=4.6
pillow>=9.2
kornia>=0.6.7
jupyter
tqdm
